Revert "[mh-11] moved dashboard tests to own file, added test for presence / absence of advance_template_directive based on user's state (which determines whether the "Download a Template" button should be shown)"
This reverts commit 468ad4a07e01cd5923b323fa00795de2be6f8496.
