+++ myhpom/tests/test_next_steps_view.py
@@ -1,7 +1,6 @@
 from django.core.urlresolvers import reverse
 from django.test import TestCase
 from django.core.files.uploadedfile import SimpleUploadedFile
-from django.db.models.fields.files import FieldFile
 
 from myhpom.models import State
 from myhpom.tests.factories import UserFactory
@@ -57,3 +56,21 @@ class NextStepsTestCase(TestCase):
             response.context['ad_template'],
         )
 
+
+class DashboardTestCase(TestCase):
+    def setUp(self):
+        self.url = reverse('myhpom:dashboard')
+
+    def test_not_logged_in(self):
+        # A user must be logged in to see their dashboard:
+        response = self.client.get(self.url)
+        self.assertEqual(302, response.status_code)
+
+    def test_basic_get(self):
+        user = UserFactory()
+        user.set_password('password')
+        user.save()
+        self.assertTrue(self.client.login(username=user.username, password='password'))
+        response = self.client.get(self.url)
+        self.assertEqual(200, response.status_code)
+        self.assertTemplateUsed('myhpom/dashboard.html')
